+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - Change merkle opening to constrain leaf [#162]
+- Update to `dusk-plonk` 0.9 [#165]
 
 ## [0.22.0] - 2021-07-27
 
@@ -280,6 +283,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Poseidon252 Sponge-hash impl with BulletProofs.
 - Variants of sponge for `Scalar` & `Gadget(Variable/LC)`.
 
+[#165]: https://github.com/dusk-network/poseidon252/issues/165
 [#162]: https://github.com/dusk-network/poseidon252/issues/162
 [#158]: https://github.com/dusk-network/poseidon252/issues/158
 [#153]: https://github.com/dusk-network/poseidon252/issues/153