+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.0] - 2022-08-17
+
 ### Added
 
 - Add `rkyv` implementation behind feature gate [#175]
@@ -322,7 +324,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#119]: https://github.com/dusk-network/poseidon252/issues/119
 [#117]: https://github.com/dusk-network/poseidon252/issues/117
 [#112]: https://github.com/dusk-network/poseidon252/issues/112
-[unreleased]: https://github.com/dusk-network/poseidon252/compare/v0.22.0...HEAD
+[unreleased]: https://github.com/dusk-network/poseidon252/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/dusk-network/poseidon252/compare/v0.22.0...v0.26.0
 [0.22.0]: https://github.com/dusk-network/poseidon252/compare/v0.20.0...v0.22.0
 [0.20.0]: https://github.com/dusk-network/poseidon252/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/dusk-network/poseidon252/compare/v0.18.0...v0.19.0